Move NavLink out to its own file and add Header unit tests
