+++ src/components/Header/NavLink/index.ts
@@ -1,2 +1,1 @@
-export * from "./Header";
 export * from "./NavLink";
